Changed ImageFormat::ScreenResolution to be std::uint32_t

+++ ImageFormats/ImageFormat.hpp
@@ -8,7 +8,7 @@
 class ImageFormat
 {
 public:
-    using ScreenResolution = std::uint16_t;
+    using ScreenResolution = std::uint32_t;
     using DataRow = std::vector<Pixel>;
     using DataStorage = std::vector<DataRow>;
     // [Height, Width]
